+++ build.gradle
@@ -157,11 +157,21 @@ sourceSets {
     compileClasspath += main.output + test.output
     runtimeClasspath += main.output + test.output
   }
+  integrationSources {
+    java.srcDir 'src/test/integration'
+    resources.srcDir 'src/test/resources'
+    compileClasspath += main.output + test.output
+    runtimeClasspath += main.output + test.output
+  }
 }
+
+test.maxParallelForks = 1 //Runtime.runtime.availableProcessors().intdiv(2) ?: 1
 
 configurations {
   functionalCompile.extendsFrom testCompile
   functionalRuntime.extendsFrom testRuntime
+  integrationSourcesCompile.extendsFrom testCompile
+  integrationSourcesRuntime.extendsFrom testRuntime
 }
 
 task smoke(type: Test, description: 'Runs the smoke tests.', group: 'Verification') {
@@ -177,3 +187,24 @@ task functional(type: Test, description: 'Runs the functional tests.', group: 'V
   testClassesDirs = sourceSets.functional.output.classesDirs
   classpath = sourceSets.functional.runtimeClasspath
 }
+
+task integrationFakeEvents(type: Test, description: 'Runs the integration tests.', group: 'Verification') {
+  testClassesDirs = sourceSets.integrationSources.output.classesDirs
+  classpath = sourceSets.integrationSources.runtimeClasspath
+
+  useJUnit {
+    includeCategories 'uk.gov.hmcts.reform.sandl.snlapi.fakeevents.IntegrationTestWithFakeBackend'
+  }
+}
+
+task integrationRealEvents(type: Test, description: 'Runs the integration tests.', group: 'Verification') {
+  testClassesDirs = sourceSets.integrationSources.output.classesDirs
+  classpath = sourceSets.integrationSources.runtimeClasspath
+
+  useJUnit {
+    includeCategories 'uk.gov.hmcts.reform.sandl.snlapi.realevents.IntegrationTestWithRealBackend'
+  }
+}
+
+test.finalizedBy integrationFakeEvents
+
