+++ src/test/integration/uk/gov/hmcts/reform/sandl/snlapi/IntegrationTest.java
@@ -1,0 +1,4 @@
+package uk.gov.hmcts.reform.sandl.snlapi;
+
+public interface IntegrationTest {
+}
