+++ src/test/integration/uk/gov/hmcts/reform/sandl/snlapi/fakeevents/IntegrationTestWithFakeBackend.java
@@ -1,0 +1,4 @@
+package uk.gov.hmcts.reform.sandl.snlapi.fakeevents;
+
+public interface IntegrationTestWithFakeBackend {
+}
