+++ src/test/integration/uk/gov/hmcts/reform/sandl/snlapi/fakeevents/controllers/SearchControllerIntTest.java
@@ -1,8 +1,8 @@
-package integration.fakeevents.controllers;
+package uk.gov.hmcts.reform.sandl.snlapi.fakeevents.controllers;
 
-import integration.fakeevents.BaseIntegrationTestWithFakeEvents;
 import io.restassured.http.Header;
 import org.junit.Test;
+import uk.gov.hmcts.reform.sandl.snlapi.fakeevents.BaseIntegrationTestWithFakeEvents;
 
 import static com.github.tomakehurst.wiremock.client.WireMock.aResponse;
 import static com.github.tomakehurst.wiremock.client.WireMock.get;
@@ -12,28 +12,28 @@ import static com.github.tomakehurst.wiremock.client.WireMock.urlEqualTo;
 import static com.github.tomakehurst.wiremock.client.WireMock.verify;
 import static io.restassured.RestAssured.given;
 
-public class RoomControllerIntTest extends BaseIntegrationTestWithFakeEvents {
+public class SearchControllerIntTest extends BaseIntegrationTestWithFakeEvents {
 
     @Test
-    public void rooms_shouldCallProperEventsEndpointAndReturnRooms() throws Exception {
-        stubFor(get(urlEqualTo("/room"))
+    public void search_shouldCallProperEventsEndpointAndReturnResults() throws Exception {
+        stubFor(get(urlEqualTo("/search?from=1&to=2&durationInSeconds=3"))
             .willReturn(aResponse()
                 .withStatus(200)
                 .withHeader("Content-Type", "application/json")
-                .withBody("[\"rooms\"]")));
+                .withBody("[\"results\"]")));
 
         Header authenticationHeader = signIn("officer1", "asd");
 
         given()
             .contentType("application/json")
             .header(authenticationHeader)
-            .when()
-            .get("/room")
-            .then()
+        .when()
+            .get("/search?from=1&to=2&durationInSeconds=3")
+        .then()
             .assertThat()
             .statusCode(200)
-            .and().extract().response();
+            .and();
 
-        verify(getRequestedFor(urlEqualTo("/room")));
+        verify(getRequestedFor(urlEqualTo("/search?from=1&to=2&durationInSeconds=3")));
     }
 }
