+++ src/test/integration/uk/gov/hmcts/reform/sandl/snlapi/realevents/IntegrationTestWithRealBackend.java
@@ -1,0 +1,4 @@
+package uk.gov.hmcts.reform.sandl.snlapi.realevents;
+
+public interface IntegrationTestWithRealBackend {
+}
